- IDE0060: 未使用パラメータ警告17件をディスカードパラメータ(_)で修正
  - IDE0059: 不要な値代入警告3件をパラメータ検証ロジック追加で修正
  - CA1305: ロケール固有警告2件をCultureInfo.InvariantCulture使用で修正
  - xUnit1004: スキップテスト警告14件をAvaloniaTestBase継承で修正

  根本的な対応により警告抑制ではなく適切なコード改善を実現:
  - エラー処理とパラメータ検証の強化
  - UIテストインフラの適切な活用
  - 国際化対応の向上

+++ Baketa.Infrastructure.Platform/Windows/NativeMethods/User32Methods.cs
@@ -158,7 +158,7 @@ internal static class User32Methods
     internal const uint MONITORINFOF_PRIMARY = 0x00000001;
     
     // 特殊ウィンドウハンドル
-    internal static readonly IntPtr HWND_MESSAGE = new IntPtr(-3);
+    internal static readonly IntPtr HWND_MESSAGE = new(-3);
 }
 
 [Flags]
